A small update
Added support for replacing the voiceover for "Big Johninator".
A new button has been added to the Audio Settings to enable and disable additional voiceovers (Intermisions after 3-2, 6-2. And 2-S)
Fixed the lack of translation in the hud of fishing mini-games, as well as other minor issues in 5-S.
Fixed the inability to replace the "EXIT" texture at level 7-2.

(Again, attempts to optimize the code)